digital-lobster-builder/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "digital-lobster-builder"
version = "0.1.0"
description = "Multi-agent pipeline that transforms WordPress exports into Astro JS 5 websites"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "gradient>=0.1.0",
    "pydo>=0.3.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
